Conversion to QT6

+++ liteidex/src/utils/folderview/dirsortfilterproxymodel.cpp
@@ -24,7 +24,7 @@
 #include "dirsortfilterproxymodel.h"
 #include <QFileSystemModel>
 #include <QFileInfo>
-#include <QDirModel>
+#include <QFileSystemModel>
 #include <QDateTime>
 //lite_memory_check_begin
 #if defined(WIN32) && defined(_MSC_VER) &&  defined(_DEBUG)
@@ -73,7 +73,11 @@ bool QDirSortItemComparator::sort(const QDirSortItem &n1, const QDirSortItem &n2
         r = f1->item.lastModified().secsTo(f2->item.lastModified());
         break;
       case QDir::Size:
-          r = int(qBound<qint64>(-1, f2->item.size() - f1->item.size(), 1));
+          r = int(qBound<qint64>(
+    			static_cast<qint64>(-1),
+    			static_cast<qint64>(f2->item.size() - f1->item.size()),
+    			static_cast<qint64>(1)
+				));
         break;
       case QDir::Type:
       {
